+++ CHANGELOG.md
@@ -7,5 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2]
+
+### Added:
+- One fetch active at a time, every call to the function receives the same result
+
+### Fixed:
+- Sending form details properly
+
+## [0.0.1]
+
 ### Added:
 - Adding the ability to cache and fetch client credentials